feat(pvp): replace CreateChallenge form with NewChallengePrompt
NewChallengePrompt is the v2 race-first entry point: a settings
preview + 'Start Race' button that navigates to /pvp/race?new=1.

Deletes CreateChallenge.tsx (form-based) and UserSearch.tsx
(direct-challenge user picker) — both out of scope for v2.
