+++ renderer/src/components/PvP/index.ts
@@ -1,5 +1,4 @@
 export { default as PvPHub } from "./PvPHub";
-export { default as ChallengeCard } from "./ChallengeCard";
-export { default as CreateChallenge } from "./CreateChallenge";
-export { default as UserSearch } from "./UserSearch";
 export { default as PvPMatch } from "./PvPMatch";
+export { default as ChallengeCard } from "./ChallengeCard";
+export { default as NewChallengePrompt } from "./NewChallengePrompt";
